requirements.txt
fastapi
python-dotenv
uvicorn[standard]
pydantic
openai
google-genai
pytest
httpx
openenv-core
websockets
